Create Question-3.cpp
Consider the following program and fill in the blanks at LINE-1 to define a function pointer type Fun_Ptr such that it can be used to create a pointer to point add() function. At LINE-2 create function pointer fp to point to function add() such that it satisfies  the given test cases.